feat: Added lots of improvements in usability and structure
Broke out the generate function into a separate module with its own
command to allow rendering of prepared markdown letters.

Added output_directory config option to control where letters are saved.

Added a backup directory in the cache to store intermediate versions of
the letter and generated html.

Added a warning in the config module if the current config is invalid.
Also made it so that you can view the current config even if it is
invalid.

Moved the prompts into text files in the assets folder. Restructured the
prompts to have a clear division between developer instructions and user
requests. Rewrote the prompts a bit to improve them. Made the user
request a mako template. Added options for the user to provide their own
prompts to override the defaults.

Added option for user to specify an editor command.

Refactored the generate module to improve readability and testability.
Lots of functional decomposition.